refactor: Rename `send_response.rs` to `server_send.rs`
